final documentation touches
