+++ vortex-vector/src/lib.rs
@@ -1,21 +1,15 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+// TODO(connor): Explain what vectors are, why we need them for the new operator model of arrays,
+// differences from Arrow (builders and arrays and scalars), etc.
 //! Immutable and mutable decompressed (canonical) vectors for Vortex.
-// TODO(connor): More docs
-
-// TODO(connor):
-// - Document everything
-// - Figure out correct panic propagation
-// - Figure out exact semantics of `split_off` w.r.t. length of capacity
-// - Fix bugs in implementations
-// - Add tests
-// - Figure out error semantics on ops traits
-// - Implement PartialEq and Eq for vectors
-// - Add stubs for remaining vector variants
-// - Potentially add `TryFrom<<Type>Vector> for Vector` or some other conversion method
 
 #![deny(missing_docs)]
+#![deny(clippy::missing_docs_in_private_items)]
+#![deny(clippy::missing_errors_doc)]
+#![deny(clippy::missing_panics_doc)]
+#![deny(clippy::missing_safety_doc)]
 
 mod vector;
 pub use vector::Vector;
