+++ vortex-vector/src/primitive/vector_mut.rs
@@ -176,3 +176,69 @@ impl PTypeUpcast for PrimitiveVectorMut {
         PrimitiveVectorMut::F64(input)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::VectorOps;
+
+    #[test]
+    fn test_from_iter_with_options() {
+        // Test FromIterator<Option<T>> with different types.
+        let vec_i32: PrimitiveVectorMut =
+            PVectorMut::<i32>::from_iter(vec![Some(1), None, Some(3), None, Some(5)]).into();
+        assert_eq!(vec_i32.len(), 5);
+        let frozen = vec_i32.freeze();
+        assert_eq!(frozen.validity().true_count(), 3);
+
+        // Test empty iterator.
+        let vec_empty: PrimitiveVectorMut =
+            PVectorMut::<f64>::from_iter(std::iter::empty::<Option<f64>>()).into();
+        assert_eq!(vec_empty.len(), 0);
+
+        // Test that None values use T::default().
+        let vec_nulls: PrimitiveVectorMut = PVectorMut::<i32>::from_iter([None, None, None]).into();
+        // Check that validity is all false for nulls.
+        let frozen = vec_nulls.freeze();
+        assert_eq!(frozen.validity().true_count(), 0);
+    }
+
+    #[test]
+    fn test_from_iter_non_null() {
+        // Test FromIterator<T> for different primitive types.
+        let vec_f64: PrimitiveVectorMut =
+            PVectorMut::<f64>::from_iter([1.5, 2.5, 3.5, 4.5, 5.5]).into();
+        assert_eq!(vec_f64.len(), 5);
+        let frozen = vec_f64.freeze();
+        assert_eq!(frozen.validity().true_count(), 5); // All valid.
+
+        let vec_u16: PrimitiveVectorMut = PVectorMut::<u16>::from_iter([1u16, 2, 3, 4, 5]).into();
+        assert_eq!(vec_u16.len(), 5);
+        let frozen = vec_u16.freeze();
+        assert_eq!(frozen.validity().true_count(), 5);
+    }
+
+    #[test]
+    fn test_operations_preserve_validity() {
+        // Test split/unsplit/extend with different primitive types.
+        let mut vec: PrimitiveVectorMut =
+            PVectorMut::<i64>::from_iter([Some(100), None, Some(300), None, Some(500)]).into();
+
+        let second_half = vec.split_off(2);
+        assert_eq!(vec.len(), 2);
+        assert_eq!(second_half.len(), 3);
+
+        let first_frozen = vec.freeze();
+        let second_frozen = second_half.freeze();
+        assert_eq!(first_frozen.validity().true_count(), 1);
+        assert_eq!(second_frozen.validity().true_count(), 2);
+
+        // Test unsplit.
+        let mut vec1: PrimitiveVectorMut = PVectorMut::<u32>::from_iter([Some(1000), None]).into();
+        let vec2: PrimitiveVectorMut = PVectorMut::<u32>::from_iter([None, Some(2000)]).into();
+        vec1.unsplit(vec2);
+        assert_eq!(vec1.len(), 4);
+        let frozen = vec1.freeze();
+        assert_eq!(frozen.validity().true_count(), 2);
+    }
+}
